requirements.txt
Django==1.11.3
django-extensions==1.8.1
django-jet==1.0.6
django-model-utils==3.0.0
django-mptt==0.8.7
pytz==2017.2
six==1.10.0
